chore: Move folders and setup boilerplate files
