+++ api.planx.uk/modules/admin/session/summary.ts
@@ -6,8 +6,14 @@ import {
 import { NextFunction, Request, Response } from "express";
 import { gql } from "graphql-request";
 
-import { Breadcrumb, Flow, LowCalSession, Passport, Team } from "../../types";
-import { $api } from "../../client";
+import {
+  Breadcrumb,
+  Flow,
+  LowCalSession,
+  Passport,
+  Team,
+} from "../../../types";
+import { $api } from "../../../client";
 
 /**
  * @swagger
